changed from float to double

+++ include/ambient.h
@@ -6,7 +6,7 @@
 /*   By: cwon <cwon@student.42bangkok.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/06 18:58:47 by cwon              #+#    #+#             */
-/*   Updated: 2025/09/06 19:35:00 by cwon             ###   ########.fr       */
+/*   Updated: 2025/09/11 13:05:43 by cwon             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ typedef struct s_ambient	t_ambient;
 
 struct s_ambient
 {
-	float	ratio;
+	double	ratio;
 	t_rgb	color;
 };
