+++ include/sphere.h
@@ -6,7 +6,7 @@
 /*   By: cwon <cwon@student.42bangkok.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/06 19:06:45 by cwon              #+#    #+#             */
-/*   Updated: 2025/09/11 07:52:07 by cwon             ###   ########.fr       */
+/*   Updated: 2025/09/11 13:06:17 by cwon             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ typedef struct s_sphere	t_sphere;
 
 struct s_sphere
 {
-	float		diameter;
+	double		diameter;
 	t_rgb		color;
 	t_vector	center;
 };
